+++ persistent/ChangeLog.md
@@ -1,8 +1,30 @@
 # Changelog for persistent
 
-## 2.11.0
 
+
+## 2.11.0.0
+
+* [#1060](https://github.com/yesodweb/persistent/pull/1060)
+  * The QuasiQuoter now supports `OnDelete` and `OnUpdate` cascade options.
 * Added explicit `forall` notation to make most API functions play nice when using `TypeApplications`. (e.g. instead of `selectList @_ @_ @User [] []`, you can now write `selectList @User [] []`) [#1006](https://github.com/yesodweb/persistent/pull/1006)
+
+## 2.10.5.2
+
+* [#1041](https://github.com/yesodweb/persistent/pull/1041)
+  * Explicit foreign keys can now reference tables with custom sql name
+  * Add qualified names to the stock classes list.
+
+## 2.10.5.1
+
+* [#1024](https://github.com/yesodweb/persistent/pull/1024)
+    * Add the ability to do documentation comments in entity definition syntax. Unfortunately, TemplateHaskell cannot add documentation comments, so this can't be used to add Haddocks to entities. 
+    * Add Haddock explainers for some of the supported entity syntax in `Database.Persist.Quasi`
+
+## 2.10.5
+
+* Add the `EntityWithPrefix` type to allow users to specify a custom prefix for raw SQL queries. [#1018](https://github.com/yesodweb/persistent/pull/1018)
+* Added Acquire based API to `Database.Persist.Sql` for working with
+  connections/pools in monads which aren't MonadUnliftIO. [#984](https://github.com/yesodweb/persistent/pull/984)
 
 ## 2.10.4